refactor(bundle-items): rename CleanseBundleIncludes to BundleItems
