+++ components/index.js
@@ -9,6 +9,7 @@ export { default as Breadcrumbs } from './breadcrumbs';
 export { default as BundleCard } from './bundle-card';
 export { default as BundleCardGrid } from './bundle-card-grid';
 export { default as BundleCardSection } from './bundle-card-section';
+export { default as BundleItems } from './bundle-items';
 export { default as ButtonBar } from './button-bar';
 export { default as Card } from './card';
 export { default as CartCheckoutFooter } from './cart-checkout-footer';
@@ -22,7 +23,6 @@ export { default as CartSummary } from './cart-summary';
 export { default as CartTotals } from './cart-totals';
 export { default as CheckoutAuthentication } from './checkout-authentication';
 export { default as CheckoutItemCard } from './checkout-item-card';
-export { default as CleanseBundleIncludes } from './cleanse-bundle-includes';
 export { default as CleanseGuide } from './cleanse-guide';
 export { default as CollapsibleDrawer } from './collapsible-drawer';
 export { default as ConfirmDrawer } from './confirm-drawer';
